% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATE_est.R
\name{ATE_est}
\alias{ATE_est}
\title{ATE_est for esating average effects.}
\usage{
ATE_est(fY, fw, fA)
}
\arguments{
\item{fY}{data's outcomes.}

\item{fw}{data's weight.}

\item{fA}{data"s explore.}
}
\value{
Average effects.
}
\description{
ATE_est for esating average effects.
}
