% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOAL.R
\name{GOAL}
\alias{GOAL}
\title{GOAL is our main function for high}
\usage{
GOAL(
  data,
  var.list,
  covar,
  Trt = "Trt",
  out = "Y",
  lambda_vec,
  gamma_convergence = 10
)
}
\arguments{
\item{data}{our data.}

\item{var.list}{variable list.}

\item{covar}{covar.}

\item{Trt}{tratments.}

\item{out}{outcome.}

\item{lambda_vec}{lambda_vec.}

\item{gamma_convergence}{gamma_convergence.}
}
\value{
results.
}
\description{
GOAL is our main function for high
}
