% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOAL_CI.R
\name{GOAL_CI}
\alias{GOAL_CI}
\title{GOAL_CI for ATE.}
\usage{
GOAL_CI(
  data,
  indices,
  var.list,
  covar,
  Trt = "Trt",
  out = "Y",
  lambda_vec,
  gamma_convergence = 10
)
}
\arguments{
\item{data}{data.}

\item{indices}{indices .}

\item{var.list}{var.list.}

\item{covar}{covar .}

\item{Trt}{Trt .}

\item{out}{out.}

\item{lambda_vec}{lambda_vec.}

\item{gamma_convergence}{gamma_convergence.}
}
\value{
ATE
}
\description{
GOAL_CI for ATE.
}
