% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try.R
\name{try}
\alias{try}
\title{try formula}
\usage{
try(formula, data = data, y = FALSE, a = TRUE, b = TRUE)
}
\arguments{
\item{formula}{formula.}

\item{data}{data.}

\item{a}{x.}

\item{b}{b.}
}
\value{
y.
}
\description{
try formula
}
\examples{
a<-c(1,2,3,1,5)
b<-c(2,2,3,2,5)
data <- data.frame(a,b)
c <-try(y~a+b,data=data)
}
