% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wAMD_function.R
\name{wAMD_function}
\alias{wAMD_function}
\title{wAMD_function for easting total effects.}
\usage{
wAMD_function(DataM, varlist, trt.var, wgt, beta)
}
\arguments{
\item{DataM}{data which is uesd to match.}

\item{varlist}{variables list.}

\item{trt.var}{treatment list.}

\item{wgt}{weight .}

\item{beta}{coffectient.}
}
\value{
ret which is ...
}
\description{
wAMD_function for easting total effects.
}
